Initial commit
`app_deluge_find` makes FlexGet work with Deluge.app on macOS (instead of needing to have the deluge package and all of its dependencies installed in `site-packages`). It does this by modifying the system `$PATH` when the plugin is initialized to reference the deluge package that is contained within /Applications/Deluge.app. The contrived name of the plugin is to ensure it runs _before_ the normal Deluge plugin, so that when that plugin tries to find the deluge package, it suceeds.